requirements.txt
python-dotenv
discord-webhook
aio-mc-rcon